refactor: contact mapping in external class
