+++ dashboard/src/app/core/entities/shared/contacts.ts
@@ -1,4 +1,6 @@
-import { hasOneNotEmptyProperty } from '~/core/entities/utils';
+import { AbstractControl, FormGroup } from '@angular/forms';
+import { hasOneNotEmptyProperty, removeNullsProperties } from '~/core/entities/utils';
+import { ContactsInterface } from '../api/shared/common/interfaces/ContactsInterface';
 
 /* tslint:disable:variable-name */
 import { Contact } from './contact';
@@ -20,5 +22,22 @@ export class Contacts {
       gdpr_controller: new Contact(this.gdpr_controller).toFormValues(),
       technical: new Contact(this.technical).toFormValues(),
     };
+  }
+}
+
+export class ContactsMapper {
+  static toForm(data: Contacts): FormGroup {
+    throw new Error('Method not implemented.');
+  }
+
+  public static toModel(contactForm: AbstractControl): ContactsInterface {
+    const contacts: ContactsInterface = {};
+    if (hasOneNotEmptyProperty(contactForm.get('gdpr_controller').value))
+      contacts.gdpr_controller = removeNullsProperties(contactForm.get('gdpr_controller').value);
+    if (hasOneNotEmptyProperty(contactForm.get('technical').value))
+      contacts.technical = removeNullsProperties(contactForm.get('technical').value);
+    if (hasOneNotEmptyProperty(contactForm.get('gdpr_dpo').value))
+      contacts.gdpr_dpo = removeNullsProperties(contactForm.get('gdpr_dpo').value);
+    return contacts;
   }
 }
